[refactor] 대시보드 - 파워유저, 인기리뷰 목록 조회

+++ src/main/java/com/codeit/project/deokhugam/domain/user/dto/PowerUserDto.java
@@ -13,7 +13,7 @@ public record PowerUserDto(
     String createdAt,
     Integer rank,
     BigDecimal score,
-    Integer reviewScoreSum,
+    double reviewScoreSum,
     Integer likeCount,
     Integer commentCount
 ) {
